Refactor module structures
